requirements.txt
shiny
pandas
duckdb
chatlas
faicons
shinywidgets
plotly
ridgeplot
python-dotenv
kaleido
google-generativeai
inspect-ai
python-dotenv